Moved all of AFSS specific utils to separate file
